Accept invite view is setup for the front-end. Also midified the okok/node-api endpoint for sending invites to match the link that accepts it at the level of the front-end
If the user clicks on accept invite in the email. they are asked to auth first and after auth they are automatically redirected to the accept view. this view is to handle the sending of the bondkey to the backend and process the bonding with both users if the information decrypted at the level of the backend corresponds to an email in the db (person table)